Add ability not to follow selected item in the table
It is not always useful to always select some item (i.e. first by
sorting order, by the way it is not even how it works natively since
sort_by() applied after, and you need to call set_selected_item(0) for
this).

Consider csysdig, by default when you did not press anything it stick to
the first item by sort order, and only if you select something (Up/Down)
then it will stick to this item in the table.

And this behaviour is better, since if you did not focus any item you
want to look at the items sorted by the sort order.

And I also made the Home button to cancel this selection.